Removed nested namespaces
