+++ src/types/cart.d.ts
@@ -4,128 +4,124 @@
  * for Checkout, you can use the Checkout endpoint to convert the cart to an order.
  * DOCS: https://documentation.elasticpath.com/commerce-cloud/docs/api/carts-and-checkout/carts/index.html
  */
-import { ResourcePage } from './core'
+import { ResourcePage } from './core';
 
-export as namespace cart
+export interface CheckoutCustomer {
+  id: string
+}
 
-export namespace cart {
-  export interface CheckoutCustomer {
-    id: string
-  }
+export interface CheckoutCustomerObject {
+  email: string
+  name: string
+}
 
-  export interface CheckoutCustomerObject {
-    email: string
-    name: string
-  }
+export interface Address {
+  first_name: string
+  last_name: string
+  company_name?: string
+  line_1: string
+  line_2?: string
+  city?: string
+  postcode: string
+  county: string
+  country: string
+  phone_number?: string
+  instructions?: string
+}
 
-  export interface Address {
-    first_name: string
-    last_name: string
-    company_name?: string
-    line_1: string
-    line_2?: string
-    city?: string
-    postcode: string
-    county: string
-    country: string
-    phone_number?: string
-    instructions?: string
-  }
+export interface ItemTaxObject {
+  name: string
+  jurisdiction: string
+  code: string
+  rate: number
+}
 
-  export interface ItemTaxObject {
-    name: string
-    jurisdiction: string
-    code: string
-    rate: number
-  }
-
-  /**
-   * Core Cart Base Interface
-   * DOCS: https://documentation.elasticpath.com/commerce-cloud/docs/api/carts-and-checkout/carts/index.html
-   */
-  export interface CartBase {
-    id: string
-    type: string
-    links: {}
-    meta?: {
+/**
+ * Core Cart Base Interface
+ * DOCS: https://documentation.elasticpath.com/commerce-cloud/docs/api/carts-and-checkout/carts/index.html
+ */
+export interface CartBase {
+  id: string
+  type: string
+  links: {}
+  meta?: {
+    display_price: {
+      with_tax: {
+        amount: number
+        currency: string
+        formatted: string
+      }
       display_price: {
-        with_tax: {
+        amount: number
+        currency: string
+        formatted: string
+        tax: {
           amount: number
           currency: string
           formatted: string
         }
-        display_price: {
-          amount: number
-          currency: string
-          formatted: string
-          tax: {
-            amount: number
-            currency: string
-            formatted: string
-          }
-        }
-        timestamps: {
-          created_at: string
-          updated_at: string
-        }
+      }
+      timestamps: {
+        created_at: string
+        updated_at: string
       }
     }
   }
+}
 
-  export interface CartItem {
-    id?: string
-    tax: {}[]
-    quantity?: number
-    type: string
-  }
-
-
-  type CartInclude = 'items'
-
-  export interface CartEndpoint {
-    // CartEndpoint
-    endpoint: 'carts'
-    cartId: string
-
-    Get(): Promise<CartBase>
-
-    Items(): Promise<ResourcePage<CartItem>>
-
-    AddProduct(
-      productId: string,
-      quantity?: number,
-      data?: any
-    ): Promise<ResourcePage<CartItem>>
-
-    AddCustomItem(body: { reference: string }): Promise<ResourcePage<CartItem>>
+export interface CartItem {
+  id?: string
+  tax: {}[]
+  quantity?: number
+  type: string
+}
 
 
-    AddPromotion(code: string): Promise<ResourcePage<CartItem>>
+type CartInclude = 'items'
 
-    RemoveItem(itemId: string): Promise<ResourcePage<CartItem>>
+export interface CartEndpoint {
+  // CartEndpoint
+  endpoint: 'carts'
+  cartId: string
 
-    UpdateItem(
-      itemId: string,
-      type: string,
-      quantity: number,
-      data?: any
-    ): Promise<ResourcePage<CartItem>>
+  Get(): Promise<CartBase>
 
-    /**
-     * @deprecated Use UpdateItem method
-     */
-    UpdateItemQuantity(itemId: string, quantity: number): Promise<ResourcePage<CartItem>>
+  Items(): Promise<ResourcePage<CartItem>>
 
-    AddItemTax(itemId: string, taxData: ItemTaxObject): Promise<ResourcePage<CartItem>>
+  AddProduct(
+    productId: string,
+    quantity?: number,
+    data?: any
+  ): Promise<ResourcePage<CartItem>>
 
-    RemoveItemTax(itemId: string, taxItemId: string): Promise<ResourcePage<CartItem>>
+  AddCustomItem(body: { reference: string }): Promise<ResourcePage<CartItem>>
 
-    Checkout(
-      customer: CheckoutCustomer | CheckoutCustomerObject,
-      billingAddress: Address,
-      shippingAddress?: Address
-    ): Promise<ResourcePage<CartItem>>
 
-    Delete(): Promise<ResourcePage<CartItem>>
-  }
+  AddPromotion(code: string): Promise<ResourcePage<CartItem>>
+
+  RemoveItem(itemId: string): Promise<ResourcePage<CartItem>>
+
+  UpdateItem(
+    itemId: string,
+    type: string,
+    quantity: number,
+    data?: any
+  ): Promise<ResourcePage<CartItem>>
+
+  /**
+   * @deprecated Use UpdateItem method
+   */
+  UpdateItemQuantity(itemId: string, quantity: number): Promise<ResourcePage<CartItem>>
+
+  AddItemTax(itemId: string, taxData: ItemTaxObject): Promise<ResourcePage<CartItem>>
+
+  RemoveItemTax(itemId: string, taxItemId: string): Promise<ResourcePage<CartItem>>
+
+  Checkout(
+    customer: CheckoutCustomer | CheckoutCustomerObject,
+    billingAddress: Address,
+    shippingAddress?: Address
+  ): Promise<ResourcePage<CartItem>>
+
+  Delete(): Promise<ResourcePage<CartItem>>
 }
