--- src/types/index.d.ts
@@ -1,18 +1,0 @@
-export as namespace index
-export * from './product'
-export * from './core'
-export * from './order'
-export * from './customer'
-export * from './cart'
-export * from './address'
-export * from './inventory'
-export * from './field'
-export * from './collection'
-export * from './category'
-export * from './brand'
-export * from './currencies'
-export * from './intergration'
-export * from './job'
-export * from './file'
-export * from './flow'
-export * from './transaction'
