+++ src/types/inventory.d.ts
@@ -4,34 +4,30 @@
  * history of inventory transactions, enabling easier stock auditing.
  * DOCS: https://documentation.elasticpath.com/commerce-cloud/docs/api/catalog/inventory/index.html
  */
-import { core } from './core'
+import { QueryableResource } from './core';
 
-export as namespace inventory
+/**
+ * Core Inventory Base Interface
+ * DOCS: https://documentation.elasticpath.com/commerce-cloud/docs/api/catalog/products/index.html
+ */
+export interface InventoryBase {
+  id?: string
+  type: string
+  total: number
+  available: number
+  allocated: number
+}
 
-export namespace inventory {
-  /**
-   * Core Inventory Base Interface
-   * DOCS: https://documentation.elasticpath.com/commerce-cloud/docs/api/catalog/products/index.html
-   */
-  export interface InventoryBase {
-    id?: string
-    type: string
-    total: number
-    available: number
-    allocated: number
-  }
+/**
+ * Inventory Endpoints
+ * TODO Need to check on actually functionality
+ */
+export interface InventoryEndpoint extends QueryableResource<InventoryBase, null, null, null> {
+  endpoint: 'inventory'
 
-  /**
-   * Inventory Endpoints
-   * TODO Need to check on actually functionality
-   */
-  export interface InventoryEndpoint extends core.QueryableResource<InventoryBase, null, null, null> {
-    endpoint: 'inventory'
-
-    IncrementStock(productId: string, quantity: number): Promise<InventoryBase>
-    DecrementStock(productId: string, quantity: number): Promise<InventoryBase>
-    AllocateStock(productId: string, quantity: number): Promise<InventoryBase>
-    DeallocateStock(productId: string, quantity: number): Promise<InventoryBase>
-    GetTransactions(productId: string): Promise<InventoryBase>
-  }
+  IncrementStock(productId: string, quantity: number): Promise<InventoryBase>
+  DecrementStock(productId: string, quantity: number): Promise<InventoryBase>
+  AllocateStock(productId: string, quantity: number): Promise<InventoryBase>
+  DeallocateStock(productId: string, quantity: number): Promise<InventoryBase>
+  GetTransactions(productId: string): Promise<InventoryBase>
 }
